MA-14163 [Android Pi] LPA after recover from sleep, ANR if move mouse.
It's caued by the mAutoTimerRunnable.run() not finished before sleep.
The getPlaybackHeadPosition execute time is about [0, periond_time].
The default period time is 1s. Also the run() is called each 1 second.

To fix, calculate the positionTime in the playback thread.
Note, this will lead the play time update step based on
the media buffer time.

Change-Id: Id1341a1de3256a389ebccc53fabd0c859a24a10a
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7522
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>