MA-19601 Camera2Basic apk can't be opened anymore if been killed once
In the first launch, load the main activity, and then use the main activity
as app:defaultNavHost, navigate to "@navigation/nav_graph", in nav_graph.xml,
startDestination="@id/permissions_fragment", after granting permission,
go to selector_fragment, the first launch behave well, but on the second launch,
crash for "ID does not reference a View inside this Activity", failed to goto the
next fragment.

compared it with android11, found use the same code, apk built from android11 run well,
but apk built from android12 source code has this issue. so need to optimize the
fragmentTransaction code for android12.

AOSP website defined that "to run a FragmentTransaction, must wait until the Lifecycle
is at least STARTED", add it in fragmentTransaction, found it worked. this issue can
be fixed with adding launchWhenStarted.

Change-Id: I77b13d4f531321edd376462ba163394507bb7011
Signed-off-by: Dandan Sun <dandan.sun@.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17529
Reviewed-by: Wang Haoran <elven.wang@nxp.com>