MA-20606 Fix Camera2Basic crash error on android13
Camera2Basic app crashes, when scroll the USBCamera
video resolution page to find bottom resolution.
The root cause of the crash is that the input parameter of
the callback function 'onFling' is Null during gesture
detection triggerd, and we can not avoid the crash inside
the “onFling” function, because the exception is reported
directly by AndroidRuntime.
So covert the SwipeGestureDetector.kt to java code

Change-Id: I3499cea8563c47cd60e48eeb3edcb84c1ee32504
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19822
Reviewed-by: fang hui <hui.fang@nxp.com>