MA-20614 DirectAudioPlayer: Increment playback thread priority
On 8ulp, the LPA case sometimes meets underrun issue with log:
  AudioFlinger: pause because of UNDERRUN, framesReady = 0,minFrames = 1, mFormat = 0x1

This log means there was no data avaiable in the shared memory
between App(AudioTrack) and AudioFlinger.
It happens when system resume, AudioFlinger keeps consuming
data, but AudioTrack has no chance to produce data before data
exhausted, then underrun.

This patch increases the priority of AudioTrack, so that
it can be scheduled in time to feed data.
  (PRI:NICE) 20:0  ->  4:-16 Thread-2/com.nxp.directaudioplayer

Signed-off-by: Jindong Yue <jindong.yue@nxp.com>
Change-Id: I4ecfcceb5e7fb1abe0930a31ea0c6f827916cad4
Reviewed-on: http://androidsource.ap.freescale.net/project/19831
Reviewed-by: fang hui <hui.fang@nxp.com>