MA-21027 Fix the isp set not reserved after taking picture
Camera2Basic did not save the isp set last time, which will cause
the issue that when back to the preview after taking pictures,
the isp settings were changed to default.
In this case, the lifecycle of camera will not trigger
the onSaveInstanceState and onDestroy callback,
just save the isp settings by HashMap and restore it when
creating the view.

Change-Id: I0057eececfa7ecdd84b2eebe26add40814e63f31
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20851
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: fang hui <hui.fang@nxp.com>