MA-21721 Fix Camera2Basic automatically exit issue
The issue occurs when taking the second photo.

We use Navigation to jump between fragments,
after taking photos, navigation 'actionCameraToJpegViewer'
will be executed, when issue happened,
'currentDestination?.id' will still remains in 'image_viewer_fragment'
which should be in 'camera_fragment' normally.

Add the listener for BackPress, and use 'navigateUp' to
return to the last fragment to ensure that fragments in
stack are correct.

Changa-Id: I007348737fd948abff36831c6e711f88429a9b26
Change-Id: I51d1dd1fb79bdd6142272254f1961cbe66254e7d
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23000
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>