MA-21752 Fix the record resolution list unscrollable issue in Camera2Basic
Can not scroll the resolution list with the touch screen
when resolution numbers exceeds the current page display.

The conflict between listview scrolling and
gesture sliding caused this.

Remove the 'left' slide 'action_video_to_camera'
and the related gesture 'TouchListener';
Add the listerner for BackPress, use 'navigateUp'
to return to the last fragment;

Change-Id: I56d2dca1765fcccaa41a0183b8029192c05e6f82
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23035
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: Hui Fang <hui.fang@nxp.com>